Feat(#67): TokenService 추가 및 Token Domain 의존성 추가
