+++ application/wypl-core/build.gradle
@@ -18,6 +18,7 @@ dependencies {
     implementation project(':domain:jpa-member-domain')
     implementation project(':domain:auth-domain')
     implementation project(':domain:redis-embedded-domain')
+    implementation project(':domain:redis-token-domain')
     implementation project(':common')
 
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
