+++ domain/auth-domain/src/main/java/com/wypl/authdomain/auth/service/AuthDomainServiceImpl.java
@@ -5,7 +5,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.wypl.jpamemberdomain.member.repository.MemberRepository;
 import com.wypl.jpamemberdomain.member.repository.SocialMemberRepository;
-import com.wypl.redistokendomain.TokenRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -15,25 +14,5 @@ import lombok.RequiredArgsConstructor;
 public class AuthDomainServiceImpl {
 	private final MemberRepository memberRepository;
 	private final SocialMemberRepository socialMemberRepository;
-	private final TokenRepository tokenRepository;
 
-
-
-	public boolean checkExistsToken(String accessToken) {
-		return tokenRepository.checkExistsToken(accessToken);
-	}
-
-	@Transactional
-	public void saveToken(String accessToken, String refreshToken) {
-		tokenRepository.saveToken(accessToken, refreshToken);
-	}
-
-	public String getRefreshToken(String accessToken) {
-		return tokenRepository.getRefreshToken(accessToken);
-	}
-
-	@Transactional
-	public void deleteToken(String accessToken) {
-		tokenRepository.deleteToken(accessToken);
-	}
 }
